backend/requirements.txt
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=10.0.0

# LangChain ecosystem (using more recent versions)
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-ollama>=0.1.0

# Vector store and embeddings
chromadb>=0.4.15
sentence-transformers>=2.2.0

# Sparse retrieval
rank-bm25>=0.2.2
nltk>=3.8

# Web search and HTTP
requests>=2.31.0
tldextract>=5.0.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0

# Data handling
numpy>=1.24.0
pydantic>=2.0.0
